Move Manage Connections to the Connected Mode menu (#3384)
