+++ src/Integration.Vsix/Commands/ConnectedModeMenu/ConnectedModeMenu.vsct
@@ -8,11 +8,32 @@
       </Group>
     </Groups>
 
+    <Buttons>
+      <Button guid="ConnectedModeMenuCmdSet" id="ManageServerConnectionsCommand">
+        <Parent guid="ConnectedModeMenuCmdSet" id="ConnectedModeGroup"/>
+        <Icon guid="guidIcons" id="bmpSonarQube"/>
+        <Strings>
+          <ButtonText>Manage Server Connections...</ButtonText>
+        </Strings>
+      </Button>
+    </Buttons>
+
+    <Bitmaps>
+      <Bitmap guid="guidIcons" href="..\..\Resources\CommandIcons.png" usedList="bmpSonarQube"/>
+    </Bitmaps>
+
   </Commands>
+
   <Symbols>
     <GuidSymbol name="ConnectedModeMenuCmdSet" value="{E5D50D82-248A-4300-A2B2-B640FCC4D4FD}">
       <IDSymbol name="ConnectedModeGroup" value="0x101"/>
+      <IDSymbol name="ManageServerConnectionsCommand" value="0x102"/>
     </GuidSymbol>
+
+    <GuidSymbol name="guidIcons" value="{216049B9-855E-4091-8781-186C91F0743F}">
+      <IDSymbol name="bmpSonarQube" value="1"/>
+    </GuidSymbol>
+
     <GuidSymbol name="SonarLintMenuSet" value="{DB0701CC-1E44-41F7-97D6-29B160A70BCB}">
       <IDSymbol name="ConnectedModeMenu" value="0x1024"/>
     </GuidSymbol>
