+++ src/Integration.Vsix/Commands/PackageCommandId.cs
@@ -23,7 +23,6 @@ namespace SonarLint.VisualStudio.Integration.Vsix
     internal enum PackageCommandId
     {
         // Buttons
-        ManageConnections = 0x100,
         ProjectExcludePropertyToggle = 0x101,
         ProjectTestPropertyAuto = 0x102,
         ProjectTestPropertyTrue = 0x103,
