+++ src/Integration.Vsix/Commands/PackageCommands.vsct
@@ -34,12 +34,6 @@
     </Menus>
     
     <Buttons>
-      <Button guid="guidCmdSet" id="cmdidManageConnections" type="Button">
-        <Icon guid="guidIcons" id="bmpSonarQube"/>
-        <Strings>
-          <ButtonText>Manage SonarQube Connections...</ButtonText>
-        </Strings>
-      </Button>
       <Button guid="guidCmdSet" id="cmdidProjectExcludePropertyToggle" type="Button" priority="0x100">
         <Parent guid="guidCmdSet" id="groupidProjectProperties"/>
         <CommandFlag>DynamicVisibility</CommandFlag>
@@ -69,20 +63,10 @@
         </Strings>
       </Button>
     </Buttons>
-    
-    <Bitmaps>
-      <Bitmap guid="guidIcons" href="..\Resources\CommandIcons.png" usedList="bmpSonarQube"/>
-    </Bitmaps>
+
   </Commands>
 
   <CommandPlacements>
-    <!-- cmdidManageConnections -->
-    <CommandPlacement guid="guidCmdSet" id="cmdidManageConnections" priority="0x900">
-      <Parent guid="guidCodeAnalysisMenuGroup" id="IDG_CODEANALYSIS_TOPLEVEL"/>
-    </CommandPlacement>
-    <CommandPlacement guid="guidCmdSet" id="cmdidManageConnections" priority="0x900">
-      <Parent guid="guidCodeAnalysisMenuGroup" id="IDG_SOLUTION_ANALYZE"/>
-    </CommandPlacement>
 
     <!-- groupidProjectSonarLint -->
     <CommandPlacement guid="guidCmdSet" id="groupidProjectSonarLint" priority="0x100">
@@ -106,23 +90,12 @@
       <IDSymbol name="menuidProjectTestProperty" value="0x851"/>
       
       <!-- Commands -->
-      <IDSymbol name="cmdidManageConnections" value="0x100"/>
       <IDSymbol name="cmdidProjectExcludePropertyToggle" value="0x101"/>
       <IDSymbol name="cmdidProjectTestPropertyAutoOption" value="0x102"/>
       <IDSymbol name="cmdidProjectTestPropertyTrueOption" value="0x103"/>
       <IDSymbol name="cmdidProjectTestPropertyFalseOption" value="0x104"/>
     </GuidSymbol>
 
-    <!-- Menu command icons -->
-    <GuidSymbol name="guidIcons" value="{216049B9-855E-4091-8781-186C91F0743F}">
-      <IDSymbol name="bmpSonarQube" value="1"/>
-    </GuidSymbol>
-
-    <!-- VS Code Analysis menu/group Guid:IDs -->
-    <GuidSymbol name="guidCodeAnalysisMenuGroup" value="{FBC1A65B-A334-4EC8-8FBE-DB1C56335FB6}">
-      <IDSymbol name="IDG_CODEANALYSIS_TOPLEVEL" value="0x208"/>
-      <IDSymbol name="IDG_SOLUTION_ANALYZE" value="0x217"/>
-    </GuidSymbol>
   </Symbols>
   
 </CommandTable>
